+++ Registers.hpp
@@ -4,21 +4,22 @@
 #include <map>
 #include <string>
 #include <tuple>
+#include <stdint.h>
 #include "StructDefinition.hpp"
 
 class Registers {
 private:
 	std::map<std::string, Variable*> reg_var;
-	std::map<std::string, unsigned int> reg_const;
+	std::map<std::string, uint64_t> reg_const;
 
 public:
 	Registers();
 	virtual ~Registers();
 
 	std::tuple<bool, Variable*> getVarRegister(std::string name);
-	std::tuple<bool, unsigned int> getConstRegister(std::string name);
+	std::tuple<bool, uint64_t> getConstRegister(std::string name);
 	void addRegister(Variable *var, std::string name);
-	void addRegister(unsigned int const_value, std::string name);
+	void addRegister(uint64_t const_value, std::string name);
 };
 
 #endif /* REGISTERS_HPP_ */
